+++ NEWS.md
@@ -3,6 +3,7 @@
 * 15 palettes have been added from vangogh package. (#133)
 * 30 palettes have been added from ButterflyColors package. (#136)
 * 35 palettes have been added from MoMAColors package. (#142)
+* 32 palettes have been added from PrettyCols package. (#126)
 
 # paletteer 1.5.0